requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
asgiref
autopep8
black
click
colorama
crispy-bootstrap5
dataclasses
Django
django-crispy-forms
django-js-asset
djangorestframework
djangorestframework-simplejwt
isort
mypy-extensions
pathspec
Pillow
platformdirs
pycodestyle
PyJWT
pymongo
pytz
six
sqlparse
toml
tomli
typing_extensions
tzdata
